// src/redux/actions/types.ts
export enum types {
    AUTHENTICATION_REQUEST = 'AUTHENTICATION_REQUEST',
    AUTHENTICATION_GET_TOKEN = 'AUTHENTICATION_GET_TOKEN',
    AUTHENTICATION_SUCCESS = 'AUTHENTICATION_SUCCESS',
    AUTHENTICATION_FAILURE = 'AUTHENTICATION_FAILURE',
    AUTHENTICATION_ROLE_FAILURE = 'AUTHENTICATION_ROLE_FAILURE',

    REGISTRATION_REQUEST = 'REGISTRATION_REQUEST',
    REGISTRATION_SUCCESS = 'REGISTRATION_SUCCESS',
    REGISTRATION_FAILURE = 'REGISTRATION_FAILURE',

    USER_UPDATE_REQUEST = 'USER_UPDATE_REQUEST',
    USER_UPDATE_SUCCESS = 'USER_UPDATE_SUCCESS',
    USER_UPDATE_FAILURE = 'USER_UPDATE_FAILURE',

    LOG_OUT_REQUEST = 'LOG_OUT_REQUEST',
    LOG_OUT_SUCCESS = 'LOG_OUT_SUCCESS',
    LOG_OUT_FAILURE = 'LOG_OUT_FAILURE',

    BEGIN_API_CALL = 'BEGIN_API_CALL',
    API_CALL_ERROR = 'API_CALL_ERROR'
}